Refactor ChatEditingDocument and Cell Entries to reuse code
