+++ src/vs/workbench/contrib/chat/browser/chatEditing/chatEditingModifiedDocumentEntry.ts
@@ -3,39 +3,25 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { addDisposableListener, getWindow } from '../../../../../base/browser/dom.js';
-import { assert } from '../../../../../base/common/assert.js';
-import { DeferredPromise, RunOnceScheduler, timeout } from '../../../../../base/common/async.js';
-import { IReference, MutableDisposable, toDisposable } from '../../../../../base/common/lifecycle.js';
-import { ITransaction, autorun, observableValue, transaction } from '../../../../../base/common/observable.js';
-import { isEqual } from '../../../../../base/common/resources.js';
-import { themeColorFromId } from '../../../../../base/common/themables.js';
+import { IReference, MutableDisposable } from '../../../../../base/common/lifecycle.js';
+import { ITransaction, autorun, transaction } from '../../../../../base/common/observable.js';
 import { assertType } from '../../../../../base/common/types.js';
 import { URI } from '../../../../../base/common/uri.js';
 import { getCodeEditor } from '../../../../../editor/browser/editorBrowser.js';
-import { EditOperation, ISingleEditOperation } from '../../../../../editor/common/core/editOperation.js';
 import { StringEdit } from '../../../../../editor/common/core/edits/stringEdit.js';
-import { Range } from '../../../../../editor/common/core/range.js';
-import { IDocumentDiff, nullDocumentDiff } from '../../../../../editor/common/diff/documentDiffProvider.js';
-import { DetailedLineRangeMapping } from '../../../../../editor/common/diff/rangeMapping.js';
 import { TextEdit } from '../../../../../editor/common/languages.js';
 import { ILanguageService } from '../../../../../editor/common/languages/language.js';
-import { IModelDeltaDecoration, ITextModel, MinimapPosition, OverviewRulerLane } from '../../../../../editor/common/model.js';
+import { ITextModel } from '../../../../../editor/common/model.js';
 import { SingleModelEditStackElement } from '../../../../../editor/common/model/editStack.js';
-import { ModelDecorationOptions, createTextBufferFactoryFromSnapshot } from '../../../../../editor/common/model/textModel.js';
-import { offsetEditFromContentChanges, offsetEditFromLineRangeMapping, offsetEditToEditOperations } from '../../../../../editor/common/model/textModelStringEdit.js';
-import { IEditorWorkerService } from '../../../../../editor/common/services/editorWorker.js';
+import { createTextBufferFactoryFromSnapshot } from '../../../../../editor/common/model/textModel.js';
+import { offsetEditFromLineRangeMapping } from '../../../../../editor/common/model/textModelStringEdit.js';
 import { IModelService } from '../../../../../editor/common/services/model.js';
 import { IResolvedTextEditorModel, ITextModelService } from '../../../../../editor/common/services/resolverService.js';
-import { TextModelEditReason } from '../../../../../editor/common/textModelEditReason.js';
-import { IModelContentChangedEvent } from '../../../../../editor/common/textModelEvents.js';
 import { localize } from '../../../../../nls.js';
-import { AccessibilitySignal, IAccessibilitySignalService } from '../../../../../platform/accessibilitySignal/browser/accessibilitySignalService.js';
 import { IConfigurationService } from '../../../../../platform/configuration/common/configuration.js';
 import { IFileService } from '../../../../../platform/files/common/files.js';
 import { IInstantiationService } from '../../../../../platform/instantiation/common/instantiation.js';
 import { IMarkerService } from '../../../../../platform/markers/common/markers.js';
-import { editorSelectionBackground } from '../../../../../platform/theme/common/colorRegistry.js';
 import { IUndoRedoElement, IUndoRedoService } from '../../../../../platform/undoRedo/common/undoRedo.js';
 import { IEditorPane, SaveReason } from '../../../../common/editor.js';
 import { IFilesConfigurationService } from '../../../../services/filesConfiguration/common/filesConfigurationService.js';
@@ -44,43 +30,13 @@ import { ICellEditOperation } from '../../../notebook/common/notebookCommon.js';
 import { ChatEditKind, IModifiedFileEntry, IModifiedFileEntryEditorIntegration, ModifiedFileEntryState } from '../../common/chatEditingService.js';
 import { IChatResponseModel } from '../../common/chatModel.js';
 import { IChatService } from '../../common/chatService.js';
-import { ChatEditingCodeEditorIntegration, IDocumentDiff2 } from './chatEditingCodeEditorIntegration.js';
-import { AbstractChatEditingModifiedFileEntry, IModifiedEntryTelemetryInfo, ISnapshotEntry, pendingRewriteMinimap } from './chatEditingModifiedFileEntry.js';
+import { ChatEditingCodeEditorIntegration } from './chatEditingCodeEditorIntegration.js';
+import { AbstractChatEditingModifiedFileEntry, IModifiedEntryTelemetryInfo, ISnapshotEntry } from './chatEditingModifiedFileEntry.js';
+import { ChatEditingModifiedTextModel } from './chatEditingModifiedTextModel.js';
 import { ChatEditingSnapshotTextModelContentProvider, ChatEditingTextModelContentProvider } from './chatEditingTextModelContentProviders.js';
 
 
 export class ChatEditingModifiedDocumentEntry extends AbstractChatEditingModifiedFileEntry implements IModifiedFileEntry {
-
-	private static readonly _lastEditDecorationOptions = ModelDecorationOptions.register({
-		isWholeLine: true,
-		description: 'chat-last-edit',
-		className: 'chat-editing-last-edit-line',
-		marginClassName: 'chat-editing-last-edit',
-		overviewRuler: {
-			position: OverviewRulerLane.Full,
-			color: themeColorFromId(editorSelectionBackground)
-		},
-	});
-
-	private static readonly _pendingEditDecorationOptions = ModelDecorationOptions.register({
-		isWholeLine: true,
-		description: 'chat-pending-edit',
-		className: 'chat-editing-pending-edit',
-		minimap: {
-			position: MinimapPosition.Inline,
-			color: themeColorFromId(pendingRewriteMinimap)
-		}
-	});
-
-	private static readonly _atomicEditDecorationOptions = ModelDecorationOptions.register({
-		isWholeLine: true,
-		description: 'chat-atomic-edit',
-		className: 'chat-editing-atomic-edit',
-		minimap: {
-			position: MinimapPosition.Inline,
-			color: themeColorFromId(pendingRewriteMinimap)
-		}
-	});
 
 	readonly initialContent: string;
 
@@ -90,19 +46,13 @@ export class ChatEditingModifiedDocumentEntry extends AbstractChatEditingModifie
 	private readonly _docFileEditorModel: IResolvedTextEditorModel;
 
 	private _edit: StringEdit = StringEdit.empty;
-	private _isEditFromUs: boolean = false;
-	private _allEditsAreFromUs: boolean = true;
-	private _diffOperation: Promise<IDocumentDiff | undefined> | undefined;
-	private _diffOperationIds: number = 0;
 
-	private readonly _diffInfo = observableValue<IDocumentDiff>(this, nullDocumentDiff);
-
-	readonly changesCount = this._diffInfo.map(diff => diff.changes.length);
-
-	private readonly _editDecorationClear = this._register(new RunOnceScheduler(() => { this._editDecorations = this.modifiedModel.deltaDecorations(this._editDecorations, []); }, 500));
-	private _editDecorations: string[] = [];
+	override get changesCount() {
+		return this.modifiedTextModel.diffInfo.map(diff => diff.changes.length);
+	}
 
 	readonly originalURI: URI;
+	private readonly modifiedTextModel: ChatEditingModifiedTextModel;
 
 	constructor(
 		resourceRef: IReference<IResolvedTextEditorModel>,
@@ -117,12 +67,10 @@ export class ChatEditingModifiedDocumentEntry extends AbstractChatEditingModifie
 		@IConfigurationService configService: IConfigurationService,
 		@IFilesConfigurationService fileConfigService: IFilesConfigurationService,
 		@IChatService chatService: IChatService,
-		@IEditorWorkerService private readonly _editorWorkerService: IEditorWorkerService,
 		@ITextFileService private readonly _textFileService: ITextFileService,
 		@IFileService fileService: IFileService,
 		@IUndoRedoService undoRedoService: IUndoRedoService,
-		@IInstantiationService instantiationService: IInstantiationService,
-		@IAccessibilitySignalService private readonly _accessibilitySignalService: IAccessibilitySignalService,
+		@IInstantiationService instantiationService: IInstantiationService
 	) {
 		super(
 			resourceRef.object.textEditorModel.uri,
@@ -150,6 +98,28 @@ export class ChatEditingModifiedDocumentEntry extends AbstractChatEditingModifie
 			)
 		);
 
+		this.modifiedTextModel = this._register(instantiationService.createInstance(ChatEditingModifiedTextModel,
+			this.originalModel, this.modifiedModel, this._stateObs));
+
+		this._register(this.modifiedTextModel.onHunkAction(action => {
+			if (action === 'acceptedAllChanges') {
+				this._stateObs.set(ModifiedFileEntryState.Accepted, undefined);
+				this._notifyAction('accepted');
+			} else {
+				this._stateObs.set(ModifiedFileEntryState.Rejected, undefined);
+				this._notifyAction('rejected');
+			}
+		}));
+
+		this._register(autorun(r => {
+			const diffInfo = this.modifiedTextModel.diffInfo.read(r);
+			if (this.state.get() !== ModifiedFileEntryState.Modified) {
+				return;
+			}
+			// If the diff has changed, then recompute the edit
+			this._edit = offsetEditFromLineRangeMapping(this.originalModel, this.modifiedModel, diffInfo.changes);
+		}));
+
 		// Create a reference to this model to avoid it being disposed from under our nose
 		(async () => {
 			const reference = await textModelService.createModelReference(docSnapshot.uri);
@@ -161,10 +131,16 @@ export class ChatEditingModifiedDocumentEntry extends AbstractChatEditingModifie
 		})();
 
 
-		this._register(this.modifiedModel.onDidChangeContent(e => this._mirrorEdits(e)));
+		this._register(this.modifiedModel.onDidChangeContent(e => {
+			this._edit = this.modifiedTextModel.mirrorEdits(e, this._edit);
 
-		this._register(toDisposable(() => {
-			this._clearCurrentEditLineDecoration();
+			if (!this.modifiedTextModel.isEditFromUs) {
+				this._userEditScheduler.schedule();
+				const didResetToOriginalContent = this.modifiedModel.getValue() === this.initialContent;
+				if (this._stateObs.get() === ModifiedFileEntryState.Modified && didResetToOriginalContent) {
+					this._stateObs.set(ModifiedFileEntryState.Rejected, undefined);
+				}
+			}
 		}));
 
 		const resourceFilter = this._register(new MutableDisposable());
@@ -178,10 +154,6 @@ export class ChatEditingModifiedDocumentEntry extends AbstractChatEditingModifie
 				resourceFilter.clear();
 			}
 		}));
-	}
-
-	private _clearCurrentEditLineDecoration() {
-		this._editDecorations = this.modifiedModel.deltaDecorations(this._editDecorations, []);
 	}
 
 	equalsSnapshot(snapshot: ISnapshotEntry | undefined): boolean {
@@ -211,78 +183,23 @@ export class ChatEditingModifiedDocumentEntry extends AbstractChatEditingModifie
 		this._stateObs.set(snapshot.state, undefined);
 		this.originalModel.setValue(snapshot.original);
 		if (restoreToDisk) {
-			this._setDocValue(snapshot.current);
+			this.modifiedTextModel.setDocValue(snapshot.current);
 		}
 		this._edit = snapshot.originalToCurrentEdit;
-		this._updateDiffInfoSeq();
+		this.modifiedTextModel.updateDiffInfo();
 	}
 
 	resetToInitialContent() {
-		this._setDocValue(this.initialContent);
+		this.modifiedTextModel.setDocValue(this.initialContent);
 	}
 
 	protected override async _areOriginalAndModifiedIdentical(): Promise<boolean> {
-		const diff = await this._diffOperation;
-		return diff ? diff.identical : false;
+		return this.modifiedTextModel.areOriginalAndModifiedIdentical();
 	}
 
 	protected override _resetEditsState(tx: ITransaction): void {
 		super._resetEditsState(tx);
-		this._clearCurrentEditLineDecoration();
-	}
-
-	private _mirrorEdits(event: IModelContentChangedEvent) {
-		const edit = offsetEditFromContentChanges(event.changes);
-
-		if (this._isEditFromUs) {
-			const e_sum = this._edit;
-			const e_ai = edit;
-			this._edit = e_sum.compose(e_ai);
-		} else {
-
-			//           e_ai
-			//   d0 ---------------> s0
-			//   |                   |
-			//   |                   |
-			//   | e_user_r          | e_user
-			//   |                   |
-			//   |                   |
-			//   v       e_ai_r      v
-			///  d1 ---------------> s1
-			//
-			// d0 - document snapshot
-			// s0 - document
-			// e_ai - ai edits
-			// e_user - user edits
-			//
-			const e_ai = this._edit;
-			const e_user = edit;
-
-			const e_user_r = e_user.tryRebase(e_ai.inverse(this.originalModel.getValue()), true);
-
-			if (e_user_r === undefined) {
-				// user edits overlaps/conflicts with AI edits
-				this._edit = e_ai.compose(e_user);
-			} else {
-				const edits = offsetEditToEditOperations(e_user_r, this.originalModel);
-				this.originalModel.applyEdits(edits);
-				this._edit = e_ai.tryRebase(e_user_r);
-			}
-
-			this._allEditsAreFromUs = false;
-			this._userEditScheduler.schedule();
-			this._updateDiffInfoSeq();
-
-			const didResetToOriginalContent = this.modifiedModel.getValue() === this.initialContent;
-			const currentState = this._stateObs.get();
-			switch (currentState) {
-				case ModifiedFileEntryState.Modified:
-					if (didResetToOriginalContent) {
-						this._stateObs.set(ModifiedFileEntryState.Rejected, undefined);
-						break;
-					}
-			}
-		}
+		this.modifiedTextModel.clearCurrentEditLineDecoration();
 	}
 
 	protected override _createUndoRedoElement(response: IChatResponseModel): IUndoRedoElement {
@@ -293,75 +210,7 @@ export class ChatEditingModifiedDocumentEntry extends AbstractChatEditingModifie
 
 	async acceptAgentEdits(resource: URI, textEdits: (TextEdit | ICellEditOperation)[], isLastEdits: boolean, responseModel: IChatResponseModel): Promise<void> {
 
-		assertType(textEdits.every(TextEdit.isTextEdit), 'INVALID args, can only handle text edits');
-		assert(isEqual(resource, this.modifiedURI), ' INVALID args, can only edit THIS document');
-
-		const isAtomicEdits = textEdits.length > 0 && isLastEdits;
-
-		let rewriteRatio = 0;
-
-		if (isAtomicEdits) {
-			// EDIT and DONE
-			const minimalEdits = await this._editorWorkerService.computeMoreMinimalEdits(this.modifiedModel.uri, textEdits) ?? textEdits;
-			const ops = minimalEdits.map(TextEdit.asEditOperation);
-			const undoEdits = this._applyEdits(ops);
-
-			if (undoEdits.length > 0) {
-				let range: Range | undefined;
-				for (let i = 0; i < undoEdits.length; i++) {
-					const op = undoEdits[i];
-					if (!range) {
-						range = Range.lift(op.range);
-					} else {
-						range = Range.plusRange(range, op.range);
-					}
-				}
-				if (range) {
-
-					const defer = new DeferredPromise<void>();
-					const listener = addDisposableListener(getWindow(undefined), 'animationend', e => {
-						if (e.animationName === 'kf-chat-editing-atomic-edit') { // CHECK with chat.css
-							defer.complete();
-							listener.dispose();
-						}
-					});
-
-					this._editDecorations = this.modifiedModel.deltaDecorations(this._editDecorations, [{
-						options: ChatEditingModifiedDocumentEntry._atomicEditDecorationOptions,
-						range
-					}]);
-
-					await Promise.any([defer.p, timeout(500)]); // wait for animation to finish but also time-cap it
-					listener.dispose();
-				}
-			}
-
-
-		} else {
-			// EDIT a bit, then DONE
-			const ops = textEdits.map(TextEdit.asEditOperation);
-			const undoEdits = this._applyEdits(ops);
-			const maxLineNumber = undoEdits.reduce((max, op) => Math.max(max, op.range.startLineNumber), 0);
-			rewriteRatio = Math.min(1, maxLineNumber / this.modifiedModel.getLineCount());
-
-			const newDecorations: IModelDeltaDecoration[] = [
-				// decorate pending edit (region)
-				{
-					options: ChatEditingModifiedDocumentEntry._pendingEditDecorationOptions,
-					range: new Range(maxLineNumber + 1, 1, Number.MAX_SAFE_INTEGER, Number.MAX_SAFE_INTEGER)
-				}
-			];
-
-			if (maxLineNumber > 0) {
-				// decorate last edit
-				newDecorations.push({
-					options: ChatEditingModifiedDocumentEntry._lastEditDecorationOptions,
-					range: new Range(maxLineNumber, 1, maxLineNumber, Number.MAX_SAFE_INTEGER)
-				});
-			}
-			this._editDecorations = this.modifiedModel.deltaDecorations(this._editDecorations, newDecorations);
-
-		}
+		const result = await this.modifiedTextModel.acceptAgentEdits(resource, textEdits, isLastEdits, responseModel);
 
 		transaction((tx) => {
 			this._waitsForLastEdits.set(!isLastEdits, tx);
@@ -369,13 +218,10 @@ export class ChatEditingModifiedDocumentEntry extends AbstractChatEditingModifie
 
 			if (!isLastEdits) {
 				this._isCurrentlyBeingModifiedByObs.set(responseModel, tx);
-				this._rewriteRatioObs.set(rewriteRatio, tx);
-
+				this._rewriteRatioObs.set(result.rewriteRatio, tx);
 			} else {
 				this._resetEditsState(tx);
-				this._updateDiffInfoSeq();
 				this._rewriteRatioObs.set(1, tx);
-				this._editDecorationClear.schedule();
 			}
 		});
 		if (isLastEdits) {
@@ -386,114 +232,10 @@ export class ChatEditingModifiedDocumentEntry extends AbstractChatEditingModifie
 		}
 	}
 
-	private async _acceptHunk(change: DetailedLineRangeMapping): Promise<boolean> {
-		if (!this._diffInfo.get().changes.includes(change)) {
-			// diffInfo should have model version ids and check them (instead of the caller doing that)
-			return false;
-		}
-		const edits: ISingleEditOperation[] = [];
-		for (const edit of change.innerChanges ?? []) {
-			const newText = this.modifiedModel.getValueInRange(edit.modifiedRange);
-			edits.push(EditOperation.replace(edit.originalRange, newText));
-		}
-		this.originalModel.pushEditOperations(null, edits, _ => null);
-		await this._updateDiffInfoSeq();
-		if (this._diffInfo.get().identical) {
-			this._stateObs.set(ModifiedFileEntryState.Accepted, undefined);
-			this._notifyAction('accepted');
-		}
-		this._accessibilitySignalService.playSignal(AccessibilitySignal.editsKept, { allowManyInParallel: true });
-		return true;
-	}
-
-	private async _rejectHunk(change: DetailedLineRangeMapping): Promise<boolean> {
-		if (!this._diffInfo.get().changes.includes(change)) {
-			return false;
-		}
-		const edits: ISingleEditOperation[] = [];
-		for (const edit of change.innerChanges ?? []) {
-			const newText = this.originalModel.getValueInRange(edit.originalRange);
-			edits.push(EditOperation.replace(edit.modifiedRange, newText));
-		}
-		this.modifiedModel.pushEditOperations(null, edits, _ => null);
-		await this._updateDiffInfoSeq();
-		if (this._diffInfo.get().identical) {
-			this._stateObs.set(ModifiedFileEntryState.Rejected, undefined);
-			this._notifyAction('rejected');
-		}
-		this._accessibilitySignalService.playSignal(AccessibilitySignal.editsUndone, { allowManyInParallel: true });
-		return true;
-	}
-
-	private _applyEdits(edits: ISingleEditOperation[]) {
-		// make the actual edit
-		this._isEditFromUs = true;
-		try {
-			let result: ISingleEditOperation[] = [];
-			TextModelEditReason.editWithReason(new TextModelEditReason({ source: 'Chat.applyEdits' }), () => {
-				this.modifiedModel.pushEditOperations(null, edits, (undoEdits) => {
-					result = undoEdits;
-					return null;
-				});
-			});
-			return result;
-		} finally {
-			this._isEditFromUs = false;
-		}
-	}
-
-	private async _updateDiffInfoSeq() {
-		const myDiffOperationId = ++this._diffOperationIds;
-		await Promise.resolve(this._diffOperation);
-		if (this._diffOperationIds === myDiffOperationId) {
-			const thisDiffOperation = this._updateDiffInfo();
-			this._diffOperation = thisDiffOperation;
-			await thisDiffOperation;
-		}
-	}
-
-	private async _updateDiffInfo(): Promise<IDocumentDiff | undefined> {
-
-		if (this.originalModel.isDisposed() || this.modifiedModel.isDisposed()) {
-			return undefined;
-		}
-
-		if (this.state.get() !== ModifiedFileEntryState.Modified) {
-			this._diffInfo.set(nullDocumentDiff, undefined);
-			return nullDocumentDiff;
-		}
-
-		const docVersionNow = this.modifiedModel.getVersionId();
-		const snapshotVersionNow = this.originalModel.getVersionId();
-
-		const diff = await this._editorWorkerService.computeDiff(
-			this.originalModel.uri,
-			this.modifiedModel.uri,
-			{
-				ignoreTrimWhitespace: false, // NEVER ignore whitespace so that undo/accept edits are correct and so that all changes (1 of 2) are spelled out
-				computeMoves: false,
-				maxComputationTimeMs: 3000
-			},
-			'advanced'
-		);
-
-		if (this.originalModel.isDisposed() || this.modifiedModel.isDisposed()) {
-			return undefined;
-		}
-
-		// only update the diff if the documents didn't change in the meantime
-		if (this.modifiedModel.getVersionId() === docVersionNow && this.originalModel.getVersionId() === snapshotVersionNow) {
-			const diff2 = diff ?? nullDocumentDiff;
-			this._diffInfo.set(diff2, undefined);
-			this._edit = offsetEditFromLineRangeMapping(this.originalModel, this.modifiedModel, diff2.changes);
-			return diff2;
-		}
-		return undefined;
-	}
 
 	protected override async _doAccept(): Promise<void> {
 		this.originalModel.setValue(this.modifiedModel.createSnapshot());
-		this._diffInfo.set(nullDocumentDiff, undefined);
+		await this.modifiedTextModel.updateDiffInfo();
 		this._edit = StringEdit.empty;
 		this._multiDiffEntryDelegate.collapse(undefined);
 
@@ -521,8 +263,8 @@ export class ChatEditingModifiedDocumentEntry extends AbstractChatEditingModifie
 			}
 			this._onDidDelete.fire();
 		} else {
-			this._setDocValue(this.originalModel.getValue());
-			if (this._allEditsAreFromUs && isTextFileEditorModel(this._docFileEditorModel)) {
+			this.modifiedTextModel.setDocValue(this.originalModel.getValue());
+			if (this.modifiedTextModel.allEditsAreFromUs && isTextFileEditorModel(this._docFileEditorModel)) {
 				// save the file after discarding so that the dirty indicator goes away
 				// and so that an intermediate saved state gets reverted
 				await this._docFileEditorModel.save({ reason: SaveReason.EXPLICIT, skipSaveParticipants: true });
@@ -531,32 +273,11 @@ export class ChatEditingModifiedDocumentEntry extends AbstractChatEditingModifie
 		}
 	}
 
-	private _setDocValue(value: string): void {
-		if (this.modifiedModel.getValue() !== value) {
-
-			this.modifiedModel.pushStackElement();
-			const edit = EditOperation.replace(this.modifiedModel.getFullModelRange(), value);
-
-			this._applyEdits([edit]);
-			this._updateDiffInfoSeq();
-			this.modifiedModel.pushStackElement();
-		}
-	}
-
-
 	protected _createEditorIntegration(editor: IEditorPane): IModifiedFileEntryEditorIntegration {
 		const codeEditor = getCodeEditor(editor.getControl());
 		assertType(codeEditor);
 
-		const diffInfo = this._diffInfo.map(value => {
-			return {
-				...value,
-				originalModel: this.originalModel,
-				modifiedModel: this.modifiedModel,
-				keep: changes => this._acceptHunk(changes),
-				undo: changes => this._rejectHunk(changes)
-			} satisfies IDocumentDiff2;
-		});
+		const diffInfo = this.modifiedTextModel.diffInfo;
 
 		return this._instantiationService.createInstance(ChatEditingCodeEditorIntegration, this, codeEditor, diffInfo, false);
 	}
