+++ src/vs/workbench/contrib/chat/browser/chatEditing/notebook/chatEditingNotebookCellEntry.ts
@@ -3,32 +3,23 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { RunOnceScheduler } from '../../../../../../base/common/async.js';
-import { Disposable, DisposableStore, toDisposable } from '../../../../../../base/common/lifecycle.js';
-import { ITransaction, IObservable, observableValue, autorun, transaction } from '../../../../../../base/common/observable.js';
-import { themeColorFromId } from '../../../../../../base/common/themables.js';
+import { Disposable, DisposableStore } from '../../../../../../base/common/lifecycle.js';
+import { IObservable, observableValue, transaction, autorun } from '../../../../../../base/common/observable.js';
 import { URI } from '../../../../../../base/common/uri.js';
-import { EditOperation, ISingleEditOperation } from '../../../../../../editor/common/core/editOperation.js';
 import { StringEdit } from '../../../../../../editor/common/core/edits/stringEdit.js';
-import { Range } from '../../../../../../editor/common/core/range.js';
-import { IDocumentDiff, nullDocumentDiff } from '../../../../../../editor/common/diff/documentDiffProvider.js';
+import { IDocumentDiff } from '../../../../../../editor/common/diff/documentDiffProvider.js';
 import { DetailedLineRangeMapping } from '../../../../../../editor/common/diff/rangeMapping.js';
 import { TextEdit } from '../../../../../../editor/common/languages.js';
-import { IModelDeltaDecoration, ITextModel, MinimapPosition, OverviewRulerLane } from '../../../../../../editor/common/model.js';
-import { ModelDecorationOptions } from '../../../../../../editor/common/model/textModel.js';
-import { offsetEditFromContentChanges, offsetEditFromLineRangeMapping, offsetEditToEditOperations } from '../../../../../../editor/common/model/textModelStringEdit.js';
-import { IEditorWorkerService } from '../../../../../../editor/common/services/editorWorker.js';
-import { IModelContentChangedEvent } from '../../../../../../editor/common/textModelEvents.js';
-import { IConfigurationService } from '../../../../../../platform/configuration/common/configuration.js';
-import { observableConfigValue } from '../../../../../../platform/observable/common/platformObservableUtils.js';
-import { editorSelectionBackground } from '../../../../../../platform/theme/common/colorRegistry.js';
+import { ITextModel } from '../../../../../../editor/common/model.js';
+import { offsetEditFromLineRangeMapping } from '../../../../../../editor/common/model/textModelStringEdit.js';
+import { IInstantiationService } from '../../../../../../platform/instantiation/common/instantiation.js';
 import { CellEditState } from '../../../../notebook/browser/notebookBrowser.js';
 import { INotebookEditorService } from '../../../../notebook/browser/services/notebookEditorService.js';
 import { NotebookCellTextModel } from '../../../../notebook/common/model/notebookCellTextModel.js';
 import { CellKind } from '../../../../notebook/common/notebookCommon.js';
 import { ModifiedFileEntryState } from '../../../common/chatEditingService.js';
 import { IChatResponseModel } from '../../../common/chatModel.js';
-import { pendingRewriteMinimap } from '../chatEditingModifiedFileEntry.js';
+import { ChatEditingModifiedTextModel } from '../chatEditingModifiedTextModel.js';
 
 
 /**
@@ -37,200 +28,93 @@ import { pendingRewriteMinimap } from '../chatEditingModifiedFileEntry.js';
  * Hence most of the same functionality applies.
  */
 export class ChatEditingNotebookCellEntry extends Disposable {
-	private static readonly _lastEditDecorationOptions = ModelDecorationOptions.register({
-		isWholeLine: true,
-		description: 'chat-last-edit',
-		className: 'chat-editing-last-edit-line',
-		marginClassName: 'chat-editing-last-edit',
-		overviewRuler: {
-			position: OverviewRulerLane.Full,
-			color: themeColorFromId(editorSelectionBackground)
-		},
-	});
-
-	private static readonly _pendingEditDecorationOptions = ModelDecorationOptions.register({
-		isWholeLine: true,
-		description: 'chat-pending-edit',
-		className: 'chat-editing-pending-edit',
-		minimap: {
-			position: MinimapPosition.Inline,
-			color: themeColorFromId(pendingRewriteMinimap)
-		}
-	});
-
-
 	public get isDisposed(): boolean {
 		return this._store.isDisposed;
 	}
 
 	private _edit: StringEdit = StringEdit.empty;
-	private _isEditFromUs: boolean = false;
 	public get isEditFromUs(): boolean {
-		return this._isEditFromUs;
+		return this.modifiedTextModel.isEditFromUs;
 	}
 
-	private _allEditsAreFromUs: boolean = true;
 	public get allEditsAreFromUs(): boolean {
-		return this._allEditsAreFromUs;
+		return this.modifiedTextModel.allEditsAreFromUs;
 	}
-	private _diffOperation: Promise<any> | undefined;
-	private _diffOperationIds: number = 0;
-
-	private readonly _diffInfo = observableValue<IDocumentDiff>(this, nullDocumentDiff);
 	public get diffInfo(): IObservable<IDocumentDiff> {
-		return this._diffInfo;
+		return this.modifiedTextModel.diffInfo;
 	}
 	private readonly _maxModifiedLineNumber = observableValue<number>(this, 0);
 	readonly maxModifiedLineNumber = this._maxModifiedLineNumber;
 
-	private readonly _editDecorationClear = this._register(new RunOnceScheduler(() => { this._editDecorations = this.modifiedModel.deltaDecorations(this._editDecorations, []); }, 500));
-	private _editDecorations: string[] = [];
-
-	private readonly _diffTrimWhitespace: IObservable<boolean>;
 	protected readonly _stateObs = observableValue<ModifiedFileEntryState>(this, ModifiedFileEntryState.Modified);
 	readonly state: IObservable<ModifiedFileEntryState> = this._stateObs;
-	protected readonly _isCurrentlyBeingModifiedByObs = observableValue<IChatResponseModel | undefined>(this, undefined);
-	readonly isCurrentlyBeingModifiedBy: IObservable<IChatResponseModel | undefined> = this._isCurrentlyBeingModifiedByObs;
 	private readonly initialContent: string;
-
+	private readonly modifiedTextModel: ChatEditingModifiedTextModel;
 	constructor(
 		public readonly notebookUri: URI,
 		public readonly cell: NotebookCellTextModel,
 		private readonly modifiedModel: ITextModel,
 		private readonly originalModel: ITextModel,
 		disposables: DisposableStore,
-		@IConfigurationService configService: IConfigurationService,
-		@IEditorWorkerService private readonly _editorWorkerService: IEditorWorkerService,
-		@INotebookEditorService private readonly notebookEditorService: INotebookEditorService
+		@INotebookEditorService private readonly notebookEditorService: INotebookEditorService,
+		@IInstantiationService private readonly instantiationService: IInstantiationService
 	) {
 		super();
 		this.initialContent = this.originalModel.getValue();
 		this._register(disposables);
-		this._register(this.modifiedModel.onDidChangeContent(e => {
-			this._mirrorEdits(e);
-		}));
-		this._register(toDisposable(() => {
-			this.clearCurrentEditLineDecoration();
+		this.modifiedTextModel = this._register(this.instantiationService.createInstance(ChatEditingModifiedTextModel, this.originalModel, this.modifiedModel, this.state));
+
+		this._register(this.modifiedTextModel.onHunkAction(action => {
+			this.revertMarkdownPreviewState();
+			if (action === 'acceptedAllChanges') {
+				this._stateObs.set(ModifiedFileEntryState.Accepted, undefined);
+			} else {
+				this._stateObs.set(ModifiedFileEntryState.Rejected, undefined);
+			}
 		}));
 
-		this._diffTrimWhitespace = observableConfigValue('diffEditor.ignoreTrimWhitespace', true, configService);
 		this._register(autorun(r => {
-			this._diffTrimWhitespace.read(r);
-			this._updateDiffInfoSeq();
+			const diffInfo = this.modifiedTextModel.diffInfo.read(r);
+			if (this.state.get() !== ModifiedFileEntryState.Modified) {
+				return;
+			}
+			// If the diff has changed, then recompute the edit
+			this._edit = offsetEditFromLineRangeMapping(this.originalModel, this.modifiedModel, diffInfo.changes);
 		}));
+
+
+		this._register(this.modifiedModel.onDidChangeContent(e => {
+			this._edit = this.modifiedTextModel.mirrorEdits(e, this._edit);
+
+			if (!this.modifiedTextModel.isEditFromUs) {
+				const didResetToOriginalContent = this.modifiedModel.getValue() === this.initialContent;
+				if (this._stateObs.get() === ModifiedFileEntryState.Modified && didResetToOriginalContent) {
+					this._stateObs.set(ModifiedFileEntryState.Rejected, undefined);
+				}
+			}
+		}));
+
 	}
 
 	public clearCurrentEditLineDecoration() {
 		if (this.modifiedModel.isDisposed()) {
 			return;
 		}
-		this._editDecorations = this.modifiedModel.deltaDecorations(this._editDecorations, []);
+		this.modifiedTextModel.clearCurrentEditLineDecoration();
 	}
 
-
-	private _mirrorEdits(event: IModelContentChangedEvent) {
-		const edit = offsetEditFromContentChanges(event.changes);
-
-		if (this._isEditFromUs) {
-			const e_sum = this._edit;
-			const e_ai = edit;
-			this._edit = e_sum.compose(e_ai);
-
-		} else {
-
-			//           e_ai
-			//   d0 ---------------> s0
-			//   |                   |
-			//   |                   |
-			//   | e_user_r          | e_user
-			//   |                   |
-			//   |                   |
-			//   v       e_ai_r      v
-			///  d1 ---------------> s1
-			//
-			// d0 - document snapshot
-			// s0 - document
-			// e_ai - ai edits
-			// e_user - user edits
-			//
-			const e_ai = this._edit;
-			const e_user = edit;
-
-			const e_user_r = e_user.tryRebase(e_ai.inverse(this.originalModel.getValue()), true);
-
-			if (e_user_r === undefined) {
-				// user edits overlaps/conflicts with AI edits
-				this._edit = e_ai.compose(e_user);
-			} else {
-				const edits = offsetEditToEditOperations(e_user_r, this.originalModel);
-				this.originalModel.applyEdits(edits);
-				this._edit = e_ai.tryRebase(e_user_r);
-			}
-
-			this._allEditsAreFromUs = false;
-			this._updateDiffInfoSeq();
-
-			const didResetToOriginalContent = this.modifiedModel.getValue() === this.initialContent;
-			const currentState = this._stateObs.get();
-			switch (currentState) {
-				case ModifiedFileEntryState.Modified:
-					if (didResetToOriginalContent) {
-						this._stateObs.set(ModifiedFileEntryState.Rejected, undefined);
-						break;
-					}
-			}
-
-		}
-	}
-
-	acceptAgentEdits(textEdits: TextEdit[], isLastEdits: boolean, responseModel: IChatResponseModel): void {
-		const notebookEditor = this.notebookEditorService.retrieveExistingWidgetFromURI(this.notebookUri)?.value;
-		if (notebookEditor) {
-			const vm = notebookEditor.getCellByHandle(this.cell.handle);
-			vm?.updateEditState(CellEditState.Editing, 'chatEdit');
-		}
-
-		const ops = textEdits.map(TextEdit.asEditOperation);
-		const undoEdits = this._applyEdits(ops);
-
-		const maxLineNumber = undoEdits.reduce((max, op) => Math.max(max, op.range.startLineNumber), 0);
-
-		const newDecorations: IModelDeltaDecoration[] = [
-			// decorate pending edit (region)
-			{
-				options: ChatEditingNotebookCellEntry._pendingEditDecorationOptions,
-				range: new Range(maxLineNumber + 1, 1, Number.MAX_SAFE_INTEGER, Number.MAX_SAFE_INTEGER)
-			}
-		];
-
-		if (maxLineNumber > 0) {
-			// decorate last edit
-			newDecorations.push({
-				options: ChatEditingNotebookCellEntry._lastEditDecorationOptions,
-				range: new Range(maxLineNumber, 1, maxLineNumber, Number.MAX_SAFE_INTEGER)
-			});
-		}
-
-		this._editDecorations = this.modifiedModel.deltaDecorations(this._editDecorations, newDecorations);
-
+	async acceptAgentEdits(textEdits: TextEdit[], isLastEdits: boolean, responseModel: IChatResponseModel): Promise<void> {
+		const { maxLineNumber } = await this.modifiedTextModel.acceptAgentEdits(this.modifiedModel.uri, textEdits, isLastEdits, responseModel);
 
 		transaction((tx) => {
 			if (!isLastEdits) {
 				this._stateObs.set(ModifiedFileEntryState.Modified, tx);
-				this._isCurrentlyBeingModifiedByObs.set(responseModel, tx);
 				this._maxModifiedLineNumber.set(maxLineNumber, tx);
 
 			} else {
-				this._resetEditsState(tx);
-				this._updateDiffInfoSeq();
 				this._maxModifiedLineNumber.set(0, tx);
-				this._editDecorationClear.schedule();
 			}
 		});
-	}
-
-	scheduleEditDecorations() {
-		this._editDecorationClear.schedule();
 	}
 
 	revertMarkdownPreviewState(): void {
@@ -248,119 +132,11 @@ export class ChatEditingNotebookCellEntry extends Disposable {
 		}
 	}
 
-	protected _resetEditsState(tx: ITransaction): void {
-		this._isCurrentlyBeingModifiedByObs.set(undefined, tx);
-		this._maxModifiedLineNumber.set(0, tx);
-	}
-
 	public async keep(change: DetailedLineRangeMapping): Promise<boolean> {
-		return this._acceptHunk(change);
-	}
-
-	private async _acceptHunk(change: DetailedLineRangeMapping): Promise<boolean> {
-		this._isEditFromUs = true;
-		try {
-			if (!this._diffInfo.get().changes.filter(c => c.modified.equals(change.modified) && c.original.equals(change.original)).length) {
-				// diffInfo should have model version ids and check them (instead of the caller doing that)
-				return false;
-			}
-			const edits: ISingleEditOperation[] = [];
-			for (const edit of change.innerChanges ?? []) {
-				const newText = this.modifiedModel.getValueInRange(edit.modifiedRange);
-				edits.push(EditOperation.replace(edit.originalRange, newText));
-			}
-			this.originalModel.pushEditOperations(null, edits, _ => null);
-		}
-		finally {
-			this._isEditFromUs = false;
-		}
-		await this._updateDiffInfoSeq();
-		if (this._diffInfo.get().identical) {
-			this.revertMarkdownPreviewState();
-			this._stateObs.set(ModifiedFileEntryState.Accepted, undefined);
-		}
-		return true;
+		return this.modifiedTextModel.diffInfo.get().keep(change);
 	}
 
 	public async undo(change: DetailedLineRangeMapping): Promise<boolean> {
-		return this._rejectHunk(change);
-	}
-
-	private async _rejectHunk(change: DetailedLineRangeMapping): Promise<boolean> {
-		this._isEditFromUs = true;
-		try {
-			if (!this._diffInfo.get().changes.includes(change)) {
-				return false;
-			}
-			const edits: ISingleEditOperation[] = [];
-			for (const edit of change.innerChanges ?? []) {
-				const newText = this.originalModel.getValueInRange(edit.originalRange);
-				edits.push(EditOperation.replace(edit.modifiedRange, newText));
-			}
-			this.modifiedModel.pushEditOperations(null, edits, _ => null);
-		} finally {
-			this._isEditFromUs = false;
-		}
-		await this._updateDiffInfoSeq();
-		if (this._diffInfo.get().identical) {
-			this.revertMarkdownPreviewState();
-			this._stateObs.set(ModifiedFileEntryState.Rejected, undefined);
-		}
-		return true;
-	}
-
-	private _applyEdits(edits: ISingleEditOperation[]) {
-		// make the actual edit
-		this._isEditFromUs = true;
-		try {
-			let result: ISingleEditOperation[] = [];
-			this.modifiedModel.pushEditOperations(null, edits, (undoEdits) => {
-				result = undoEdits;
-				return null;
-			});
-			return result;
-		} finally {
-			this._isEditFromUs = false;
-		}
-	}
-
-	private async _updateDiffInfoSeq() {
-		const myDiffOperationId = ++this._diffOperationIds;
-		await Promise.resolve(this._diffOperation);
-		if (this._diffOperationIds === myDiffOperationId) {
-			const thisDiffOperation = this._updateDiffInfo();
-			this._diffOperation = thisDiffOperation;
-			await thisDiffOperation;
-		}
-	}
-
-	private async _updateDiffInfo(): Promise<void> {
-
-		if (this.originalModel.isDisposed() || this.modifiedModel.isDisposed()) {
-			return;
-		}
-
-		const docVersionNow = this.modifiedModel.getVersionId();
-		const snapshotVersionNow = this.originalModel.getVersionId();
-
-		const ignoreTrimWhitespace = this._diffTrimWhitespace.get();
-
-		const diff = await this._editorWorkerService.computeDiff(
-			this.originalModel.uri,
-			this.modifiedModel.uri,
-			{ ignoreTrimWhitespace, computeMoves: false, maxComputationTimeMs: 3000 },
-			'advanced'
-		);
-
-		if (this.originalModel.isDisposed() || this.modifiedModel.isDisposed()) {
-			return;
-		}
-
-		// only update the diff if the documents didn't change in the meantime
-		if (this.modifiedModel.getVersionId() === docVersionNow && this.originalModel.getVersionId() === snapshotVersionNow) {
-			const diff2 = diff ?? nullDocumentDiff;
-			this._diffInfo.set(diff2, undefined);
-			this._edit = offsetEditFromLineRangeMapping(this.originalModel, this.modifiedModel, diff2.changes);
-		}
+		return this.modifiedTextModel.diffInfo.get().undo(change);
 	}
 }
